+++ commit/trk2dictionary/trk2dictionary.pyx
@@ -1,17 +1,21 @@
 #!python
 # cython: language_level=3, c_string_type=str, c_string_encoding=ascii, boundscheck=False, wraparound=False, profile=False
 from __future__ import print_function
-import cython
 import numpy as np
 cimport numpy as np
+from libc.stdlib cimport malloc, free
 import nibabel
-from os.path import join, exists, splitext, dirname, isdir
+from dicelib.clustering import run_clustering
+
+import os
+from os.path import join, exists, splitext, dirname, isdir, isfile
 from os import makedirs, remove
 import time
 import amico
 import pickle
 from amico.util import LOG, NOTE, WARNING, ERROR
 from pkg_resources import get_distribution
+import shutil
 
 from libcpp cimport bool
 
@@ -21,9 +25,9 @@ cdef extern from "trk2dictionary_c.cpp":
         char* filename_tractogram, int data_offset, int Nx, int Ny, int Nz, float Px, float Py, float Pz, int n_count, int n_scalars,
         int n_properties, float fiber_shiftX, float fiber_shiftY, float fiber_shiftZ, float min_seg_len, float min_fiber_len,  float max_fiber_len,
         float* ptrPEAKS, int Np, float vf_THR, int ECix, int ECiy, int ECiz,
-        float* _ptrMASK, float* ptrTDI, char* path_out, int c, double* ptrPeaksAffine,
+        float* _ptrMASK, double** ptrTDI, char* path_out, int c, double* ptrPeaksAffine,
         int nReplicas, double* ptrBlurRho, double* ptrBlurAngle, double* ptrBlurWeights, bool* ptrBlurApplyTo,
-        float* ptrTractsAffine, unsigned short ndirs, short* prtHashTable
+        float* ptrTractsAffine, short* prtHashTable, int threads_count
     ) nogil
 
 def _get_header( niiFILE ):
@@ -32,12 +36,29 @@ def _get_header( niiFILE ):
 def _get_affine( niiFILE ):
     return niiFILE.affine if nibabel.__version__ >= '2.0.0' else niiFILE.get_affine()
 
-cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filename_mask=None, do_intersect=True,
-    fiber_shift=0, min_seg_len=1e-3, min_fiber_len=0.0, max_fiber_len=250.0,
-    vf_THR=0.1, peaks_use_affine=False, flip_peaks=[False,False,False],
-    blur_spacing=0.25, blur_core_extent=0.0, blur_gauss_extent=0.0, blur_gauss_min=0.1, blur_apply_to=None,
-    filename_trk=None, gen_trk=None, TCK_ref_image=None, ndirs=32761
-    ):
+cpdef cat_function( infilename, outfilename ):
+    """ Concatenate binary file """
+    with open( outfilename, "wb" ) as outfile:
+        for fname in infilename:
+            with open( fname, 'rb' ) as inFile:
+                shutil.copyfileobj( inFile, outfile )
+                remove( fname )
+
+cpdef compute_tdi( np.uint32_t[::1] v, np.float32_t[::1] l, int nx, int ny, int nz ):
+    cdef np.float32_t [::1] tdi = np.zeros( nx*ny*nz, dtype=np.float32 )
+    cdef int i
+    for i in xrange(v.size):
+        tdi[ v[i] ] += l[i]
+    return tdi
+
+
+cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filename_mask=None,
+            do_intersect=True, fiber_shift=0, min_seg_len=1e-3, min_fiber_len=0.0, max_fiber_len=250.0,
+            vf_THR=0.1, peaks_use_affine=False, flip_peaks=[False,False,False], blur_clust_groupby=None,
+            blur_clust_thr=0, blur_spacing=0.25, blur_core_extent=0.0, blur_gauss_extent=0.0,
+            blur_gauss_min=0.1, blur_apply_to=None, TCK_ref_image=None, ndirs=500, n_threads=-1,
+            keep_temp=False, verbose=2
+            ):
     """Perform the conversion of a tractoram to the sparse data-structure internally
     used by COMMIT to perform the matrix-vector multiplications with the operator A
     during the inversion of the linear system.
@@ -109,21 +130,27 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
     blur_apply_to: array of bool
         For each input streamline, decide whether blur is applied or not to it (default : None, meaning apply to all).
 
+    blur_clust_thr: float
+        Clustering threshold used to remove redundant streamlines from the input tractogram
+
     ndirs : int
         Number of orientations on the sphere used to discretize the orientation of each
-        each segment in a streamline (default : 32761).
+        each segment in a streamline (default : 500).
 
-    filename_trk : string
-        DEPRECATED. Use filename_tractogram instead.
+    n_threads: int
+        Number of threads. If nothing is specified, the value used is the number of CPUs available
+        in the system (default : -1).
 
-    gen_trk : string
-        DEPRECATED. No tractogram will be saved any more, but the returned coefficients will account
-        for the streamlines that were pre-filtered in this function.
+    keep_temp: boolean
+        If True, the temporary files are not deleted (default : False).
+
+    verbose: int
+        Verbosity level (default : 2).
     """
 
     # check the value of ndirs
     if not amico.lut.is_valid(ndirs):
-        ERROR( 'Unsupported value for ndirs.\nNote: Supported values for ndirs are [500, 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500, 7000, 7500, 8000, 8500, 9000, 9500, 10000, 32761 (default)]' )
+        ERROR( 'Unsupported value for ndirs.\nNote: Supported values for ndirs are [500 (default), 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500, 7000, 7500, 8000, 8500, 9000, 9500, 10000, 32761]' )
 
     # check conflicts of fiber_shift
     if np.isscalar(fiber_shift) :
@@ -214,15 +241,8 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
     if max_fiber_len < min_fiber_len :
         ERROR( '"max_fiber_len" must be >= "min_fiber_len"' )
 
-    if filename_trk is None and filename_tractogram is None:
+    if filename_tractogram is None:
         ERROR( '"filename_tractogram" not defined' )
-
-    if filename_trk is not None and filename_tractogram is not None:
-        WARNING('"filename_trk" will not be considered, "filename_tractogram" will be used')
-
-    if filename_trk is not None and filename_tractogram is None:
-        filename_tractogram = filename_trk
-        WARNING('"filename_trk" parameter is deprecated, use "filename_tractogram" instead')
 
     if path_out is None:
         path_out = dirname(filename_tractogram)
@@ -232,13 +252,66 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
             ERROR( '"path_out" cannot be inferred from "filename_tractogram"' )
         path_out = join(path_out,'COMMIT')
 
-    if gen_trk is not None:
-        WARNING('"gen_trk" parameter is deprecated')
-
     # create output path
     print( f'\t- Output written to "{path_out}"' )
     if not exists( path_out ):
         makedirs( path_out )
+    
+    path_temp = join(path_out, 'temp')
+    print( f'\t- Temporary files written to "{path_temp}"' )
+    
+    if exists(path_temp):
+        shutil.rmtree(path_temp, ignore_errors=True)
+    makedirs(path_temp, exist_ok=True)
+
+    if n_threads == 0 or n_threads > 255 :
+        ERROR( 'Number of n_threads must be between 1 and 255' )
+        
+    if n_threads == -1 :
+        # Set to the number of CPUs in the system
+        try :
+            n_threads = os.cpu_count()
+        except :
+            n_threads = 1
+
+    print( f'\t- Using parallel computation with {n_threads} threads' )
+
+    if np.isscalar(blur_clust_thr):
+        blur_clust_thr = np.array( [blur_clust_thr] )
+
+    if blur_clust_thr[0]> 0:
+        LOG( '\n   * Running tractogram clustering:' )
+        print( f'\t- Input tractogram "{filename_tractogram}"' )
+        print( f'\t- Clustering threshold = {blur_clust_thr[0]}' )
+        input_hdr = nibabel.streamlines.load( filename_tractogram, lazy_load=True ).header
+        input_n_count = int(input_hdr['count'])
+
+        extension = splitext(filename_tractogram)[1]
+
+        if filename_mask is None and TCK_ref_image is None:
+            if extension == ".tck":
+                ERROR( 'TCK files do not contain information about the geometry. Use "TCK_ref_image" for that' )
+            else:
+                ERROR( 'Unknown file extension. Use "filename_mask" or "TCK_ref_image" for that' )
+
+        input_tractogram = os.path.basename(filename_tractogram)[:-4]
+        filename_out = join( path_temp, f'{input_tractogram}_clustered_thr_{float(blur_clust_thr[0])}.tck' )
+
+        if blur_clust_groupby:
+            file_assignments = join( path_temp, f'{input_tractogram}_clustered_thr_{blur_clust_thr[0]}_assignments.txt' )
+            hdr = nibabel.streamlines.load( filename_tractogram, lazy_load=True ).header
+            temp_idx = np.arange(int(hdr['count']))
+            path_streamline_idx = join( path_temp,"streamline_idx.npy")
+            np.save( path_streamline_idx, temp_idx )
+            idx_centroids = run_clustering(file_name_in=filename_tractogram, output_folder=path_temp, atlas=blur_clust_groupby,
+                            clust_thr=blur_clust_thr[0], save_assignments=file_assignments,
+                            temp_idx=path_streamline_idx, n_threads=n_threads, force=True, verbose=verbose) 
+        else:
+            idx_centroids = run_clustering(file_name_in=filename_tractogram, output_folder=path_temp, clust_thr=blur_clust_thr[0],
+                            force=True, verbose=verbose)
+        filename_tractogram = filename_out
+
+
 
     # Load data from files
     LOG( '\n   * Loading data:' )
@@ -255,12 +328,13 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
         ERROR( 'Invalid input file: only .trk and .tck are supported' )
 
     hdr = nibabel.streamlines.load( filename_tractogram, lazy_load=True ).header
+        
 
     if extension == ".trk":
         print ( f'\t\t- geometry taken from "{filename_tractogram}"' )
-        Nx = hdr['dimensions'][0]
-        Ny = hdr['dimensions'][1]
-        Nz = hdr['dimensions'][2]
+        Nx = int(hdr['dimensions'][0])
+        Ny = int(hdr['dimensions'][1])
+        Nz = int(hdr['dimensions'][2])
         Px = hdr['voxel_sizes'][0]
         Py = hdr['voxel_sizes'][1]
         Pz = hdr['voxel_sizes'][2]
@@ -295,7 +369,10 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
 
     print( f'\t\t- {Nx} x {Ny} x {Nz}' )
     print( f'\t\t- {Px:.4f} x {Py:.4f} x {Pz:.4f}' )
-    print( f'\t\t- {n_count} streamlines' )
+    if blur_clust_thr[0]> 0:
+        print( f'\t\t- {input_n_count} streamlines' )
+    else:
+        print( f'\t\t- {n_count} streamlines' )
     if Nx >= 2**16 or Nz >= 2**16 or Nz >= 2**16 :
         ERROR( 'The max dim size is 2^16 voxels' )
 
@@ -339,10 +416,14 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
     cdef float* ptrPEAKS
     cdef float [:, :, :, ::1] niiPEAKS_img
     cdef int Np
-    cdef float [:, :, ::1] niiTDI_img = np.ascontiguousarray( np.zeros((Nx,Ny,Nz),dtype=np.float32) )
-    cdef float* ptrTDI  = &niiTDI_img[0,0,0]
+    cdef double [:,:, :,::1] niiTDI_img = np.ascontiguousarray( np.zeros((n_threads,Nx,Ny,Nz),dtype=np.float64) )
+    cdef double** ptrTDI = <double**>malloc( n_threads * sizeof(double*) )
+    for i in range(n_threads):
+        ptrTDI[i] = &niiTDI_img[i,0,0,0]
+
     cdef double [:, ::1] peaksAffine
     cdef double* ptrPeaksAffine
+
     if filename_peaks is not None :
         print( '\t- EC orientations' )
         niiPEAKS = nibabel.load( filename_peaks )
@@ -379,6 +460,10 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
     # write dictionary information info file
     dictionary_info = {}
     dictionary_info['filename_tractogram'] = filename_tractogram
+    if blur_clust_thr[0]> 0:
+        idx_centroids = np.array(idx_centroids, dtype=np.uint32)
+        dictionary_info['n_count'] = input_n_count
+        dictionary_info['tractogram_centr_idx'] = idx_centroids 
     dictionary_info['TCK_ref_image'] = TCK_ref_image
     dictionary_info['path_out'] = path_out
     dictionary_info['filename_peaks'] = filename_peaks
@@ -398,6 +483,7 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
     dictionary_info['blur_sigma'] = blur_sigma
     dictionary_info['blur_apply_to'] = blur_apply_to
     dictionary_info['ndirs'] = ndirs
+    dictionary_info['n_threads'] = n_threads
     with open( join(path_out,'dictionary_info.pickle'), 'wb+' ) as dictionary_info_file:
         pickle.dump(dictionary_info, dictionary_info_file, protocol=2)
 
@@ -406,11 +492,73 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
         Nx, Ny, Nz, Px, Py, Pz, n_count, n_scalars, n_properties,
         fiber_shiftX, fiber_shiftY, fiber_shiftZ, min_seg_len, min_fiber_len, max_fiber_len,
         ptrPEAKS, Np, vf_THR, -1 if flip_peaks[0] else 1, -1 if flip_peaks[1] else 1, -1 if flip_peaks[2] else 1,
-        ptrMASK, ptrTDI, path_out, 1 if do_intersect else 0, ptrPeaksAffine,
-        nReplicas, &blurRho[0], &blurAngle[0], &blurWeights[0], &blurApplyTo[0], ptrToVOXMM, ndirs, ptrHashTable  );
+        ptrMASK, ptrTDI, path_temp, 1 if do_intersect else 0, ptrPeaksAffine,
+        nReplicas, &blurRho[0], &blurAngle[0], &blurWeights[0], &blurApplyTo[0], ptrToVOXMM, ptrHashTable, n_threads  );
     if ret == 0 :
         WARNING( 'DICTIONARY not generated' )
         return None
+
+    # Concatenate files together
+    cdef int discarded = 0
+    for j in range(n_threads-1):
+        path_IC_f = path_temp + f'/dictionary_IC_f_{j+1}.dict'
+        kept = np.fromfile( path_temp + f'/dictionary_TRK_kept_{j}.dict', dtype=np.uint8 )
+        IC_f = np.fromfile( path_temp + f'/dictionary_IC_f_{j+1}.dict', dtype=np.uint32 )
+        discarded += np.count_nonzero(kept==0)
+        IC_f -= discarded
+        IC_f_save = np.memmap( path_IC_f, dtype="uint32", mode='w+', shape=IC_f.shape )
+        IC_f_save[:] = IC_f[:]
+        IC_f_save.flush()
+        del IC_f_save
+        # np.save( path_out + f'/dictionary_IC_f_{j+1}.dict', IC_f, allow_pickle=True)
+
+    fileout = path_out + '/dictionary_TRK_kept.dict'
+    dict_list = []
+    for j in range(n_threads):
+        dict_list += [ path_temp + f'/dictionary_TRK_kept_{j}.dict' ]
+    cat_function( dict_list, fileout )
+
+    fileout = path_out + '/dictionary_TRK_norm.dict'
+    dict_list = []
+    for j in range(n_threads):
+        dict_list += [ path_temp + f'/dictionary_TRK_norm_{j}.dict' ]
+    cat_function( dict_list, fileout )
+
+    fileout = path_out + '/dictionary_TRK_len.dict'
+    dict_list = []
+    for j in range(n_threads):
+        dict_list += [ path_temp + f'/dictionary_TRK_len_{j}.dict' ]
+    cat_function( dict_list, fileout )
+
+    fileout = path_out + '/dictionary_TRK_lenTot.dict'
+    dict_list = []
+    for j in range(n_threads):
+        dict_list += [ path_temp + f'/dictionary_TRK_lenTot_{j}.dict' ]
+    cat_function( dict_list, fileout )
+
+    fileout = path_out + '/dictionary_IC_f.dict'
+    dict_list = []
+    for j in range(n_threads):
+        dict_list += [ path_temp + f'/dictionary_IC_f_{j}.dict' ]
+    cat_function( dict_list, fileout )
+
+    fileout = path_out + '/dictionary_IC_v.dict'
+    dict_list = []
+    for j in range(n_threads):
+        dict_list += [ path_temp + f'/dictionary_IC_v_{j}.dict' ]
+    cat_function( dict_list, fileout )
+
+    fileout = path_out + '/dictionary_IC_o.dict'
+    dict_list = []
+    for j in range(n_threads):
+        dict_list += [ path_temp + f'/dictionary_IC_o_{j}.dict' ]
+    cat_function( dict_list, fileout )
+
+    fileout = path_out + '/dictionary_IC_len.dict'
+    dict_list = []
+    for j in range(n_threads):
+        dict_list += [ path_temp + f'/dictionary_IC_len_{j}.dict' ]
+    cat_function( dict_list, fileout )
 
     # save TDI and MASK maps
     if TCK_ref_image is not None:
@@ -422,7 +570,14 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
     else :
         TDI_affine = np.diag( [Px, Py, Pz, 1] )
 
-    niiTDI = nibabel.Nifti1Image( niiTDI_img, TDI_affine )
+    v = np.fromfile( join(path_out, 'dictionary_IC_v.dict'),   dtype=np.uint32 )
+    l = np.fromfile( join(path_out, 'dictionary_IC_len.dict'), dtype=np.float32 )
+
+    cdef np.float32_t [::1] niiTDI_mem = np.zeros( Nx*Ny*Nz, dtype=np.float32 )
+    niiTDI_mem = compute_tdi( v, l, Nx, Ny, Nz )
+    niiTDI_img_save = np.reshape( niiTDI_mem, niiMASK.shape, order='F' )
+
+    niiTDI = nibabel.Nifti1Image( niiTDI_img_save, TDI_affine )
     niiTDI_hdr = _get_header( niiTDI )
     niiTDI_hdr['descrip'] = f'Created with COMMIT {get_distribution("dmri-commit").version}'
     nibabel.save( niiTDI, join(path_out,'dictionary_tdi.nii.gz') )
@@ -434,5 +589,9 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
     niiMASK_hdr = _get_header( niiMASK )
     niiMASK_hdr['descrip'] = niiTDI_hdr['descrip']
     nibabel.save( niiMASK, join(path_out,'dictionary_mask.nii.gz') )
+
+    if not keep_temp:
+        shutil.rmtree(path_temp)
+
 
     LOG( f'\n   [ {time.time() - tic:.1f} seconds ]' )