+++ commit/trk2dictionary/trk2dictionary_c.cpp
@@ -7,9 +7,17 @@
 #include "ProgressBar.h"
 #include <numpy/arrayobject.h>
 #include <math.h>
+#include <iostream>
+#include <thread>
+#include <numeric>
+#include <chrono>
 
+#define _FILE_OFFSET_BITS 64
 #define MAX_FIB_LEN 10000
+#define MAX_THREADS 255
 
+using namespace std;
+ProgressBar* PROGRESS = new ProgressBar();
 
 // CLASS to store the segments of one fiber
 class segKey
@@ -51,11 +59,12 @@ class segInVoxKey
     }
 };
 
+
 // global variables (to avoid passing them at each call)
-std::map<segKey,float>          FiberSegments;
-float                           FiberLen;      // length of a streamline
-float                           FiberLenTot;   // length of a streamline (considering the blur)
-std::vector< Vector<double> >   P;
+thread_local map<segKey,float>               FiberSegments;
+thread_local float                           FiberLen;      // length of a streamline
+thread_local float                           FiberLenTot;   // length of a streamline (considering the blur)
+thread_local vector< Vector<double> >        P;
 
 Vector<int>     dim;
 Vector<float>   pixdim;
@@ -64,57 +73,48 @@ float           fiberShiftXmm, fiberShiftYmm, fiberShiftZmm;
 bool            doIntersect;
 float           minSegLen, minFiberLen, maxFiberLen;
 
+// Threads variables
+vector<thread>  threads;
+vector<unsigned int>    totICSegments; 
+vector<unsigned int>    totFibers;
+unsigned int            totECVoxels = 0;
+unsigned int            totECSegments = 0;
+
+
+// --- Functions Definitions ----
 bool rayBoxIntersection( Vector<double>& origin, Vector<double>& direction, Vector<double>& vmin, Vector<double>& vmax, double & t);
-void fiberForwardModel( float fiber[3][MAX_FIB_LEN], unsigned int pts, int nReplicas, double* ptrBlurRho, double* ptrBlurAngle, double* ptrBlurWeights, bool doApplyBlur, short* ptrHashTable );
-void segmentForwardModel( const Vector<double>& P1, const Vector<double>& P2, int k, double w, short* ptrHashTable );
+void fiberForwardModel( float fiber[3][MAX_FIB_LEN], unsigned int pts, int nReplicas, double* ptrBlurRho, double* ptrBlurAngle, double* ptrBlurWeights, bool doApplyBlur, short* ptrHashTable, vector<Vector<double>>& P );
+void segmentForwardModel( const Vector<double>& P1, const Vector<double>& P2, int k, double w, short* ptrHashTable);
 unsigned int read_fiberTRK( FILE* fp, float fiber[3][MAX_FIB_LEN], int ns, int np );
 unsigned int read_fiberTCK( FILE* fp, float fiber[3][MAX_FIB_LEN] , float* toVOXMM );
 
 
-// =========================
-// Function called by CYTHON
-// =========================
+// ---------- Parallel fuction --------------
+int ICSegments( char* str_filename, int isTRK, int n_count, int nReplicas, int n_scalars, int n_properties, float* ptrToVOXMM,
+double* ptrTDI , double* ptrBlurRho, double* ptrBlurAngle, double* ptrBlurWeights, bool* ptrBlurApplyTo, short* ptrHashTable, char* path_out, 
+unsigned long long int offset, int idx, unsigned int startpos, unsigned int endpos );
+
+int ECSegments(float* ptrPEAKS, int Np, float vf_THR, int ECix, int ECiy, int ECiz,
+    double** ptrTDI, short* ptrHashTable, char* path_out, double* ptrPeaksAffine, int idx);
+
+
+
+// ===========================================
+//          Called by Cython
+// ===========================================
+
 int trk2dictionary(
     char* str_filename, int data_offset, int Nx, int Ny, int Nz, float Px, float Py, float Pz, int n_count, int n_scalars, int n_properties,
     float fiber_shiftX, float fiber_shiftY, float fiber_shiftZ, float min_seg_len, float min_fiber_len, float max_fiber_len,
     float* ptrPEAKS, int Np, float vf_THR, int ECix, int ECiy, int ECiz,
-    float* _ptrMASK, float* ptrTDI, char* path_out, int c, double* ptrPeaksAffine,
+    float* _ptrMASK, double** ptrTDI, char* path_out, int c, double* ptrPeaksAffine,
     int nReplicas, double* ptrBlurRho, double* ptrBlurAngle, double* ptrBlurWeights, bool* ptrBlurApplyTo,
-    float* ptrToVOXMM, unsigned short ndirs, short* ptrHashTable
+    float* ptrToVOXMM, short* ptrHashTable, int threads_count
 )
 {
-    /*=========================*/
-    /*     IC compartments     */
-    /*=========================*/
-    float          fiber[3][MAX_FIB_LEN];
-    float          fiberNorm;
-    unsigned int   N, totICSegments = 0, totFibers = 0, v;
-    unsigned short o;
-    unsigned char  kept;
-    std::string    filename;
-    std::string    OUTPUT_path(path_out);
-    std::map<segKey,float>::iterator it;
 
-    std::map<segInVoxKey,float> FiberNorm;
-    std::map<segInVoxKey,float>::iterator itNorm;
-    segInVoxKey         inVoxKey;
-
-    printf( "\n   \033[0;32m* Exporting IC compartments:\033[0m\n" );
-
-    int isTRK;
-    char *ext = strrchr(str_filename, '.');
-    if (strcmp(ext,".trk")==0) //for .trk file
-        isTRK = 1;
-    else if (strcmp(ext,".tck")==0)// for .tck file
-        isTRK = 0;
-    else
-        return 0;
-
-    FILE* fpTractogram = fopen(str_filename,"rb"); //open
-    if (fpTractogram == NULL) return 0;
-    fseek(fpTractogram,data_offset,SEEK_SET); //skip header
-
-    // set global variables
+    // Set the global variables
+    // ----------------------------------
     dim.Set( Nx, Ny, Nz );
     pixdim.Set( Px, Py, Pz );
     fiberShiftXmm = fiber_shiftX * pixdim.x; // shift in mm for the coordinates
@@ -125,93 +125,156 @@ int trk2dictionary(
     minSegLen     = min_seg_len;
     minFiberLen   = min_fiber_len;
     maxFiberLen   = max_fiber_len;
+    totICSegments.resize( threads_count, 0 );
+    totFibers.resize( threads_count, 0 );
+    totECVoxels   = 0;
+    totECSegments = 0;
 
-    P.resize( nReplicas );
 
-    // open files
-    filename = OUTPUT_path+"/dictionary_TRK_norm.dict";   FILE* pDict_TRK_norm = fopen(filename.c_str(),"wb");
-    if ( !pDict_TRK_norm )
+    // Compute the batch size for each thread
+    // ---------------------------------------
+    int p_size = n_count > threads_count ? threads_count : n_count;
+    int* batch_size = new int[p_size]();
+
+    for (int i = 0; i < n_count; i++)
     {
-        printf( "\n[trk2dictionary] Unable to create output files" );
-        return 0;
+        batch_size[i%p_size] += 1;
     }
-    filename = OUTPUT_path+"/dictionary_IC_f.dict";        FILE* pDict_IC_f       = fopen(filename.c_str(),"wb");
-    filename = OUTPUT_path+"/dictionary_IC_v.dict";        FILE* pDict_IC_v       = fopen(filename.c_str(),"wb");
-    filename = OUTPUT_path+"/dictionary_IC_o.dict";        FILE* pDict_IC_o       = fopen(filename.c_str(),"wb");
-    filename = OUTPUT_path+"/dictionary_IC_len.dict";      FILE* pDict_IC_len     = fopen(filename.c_str(),"wb");
-    filename = OUTPUT_path+"/dictionary_TRK_len.dict";     FILE* pDict_TRK_len    = fopen(filename.c_str(),"wb");
-    filename = OUTPUT_path+"/dictionary_TRK_lenTot.dict";  FILE* pDict_TRK_lenTot = fopen(filename.c_str(),"wb");
-    filename = OUTPUT_path+"/dictionary_TRK_kept.dict";    FILE* pDict_TRK_kept   = fopen(filename.c_str(),"wb");
 
-    // iterate over streamlines
-    ProgressBar PROGRESS( n_count );
-    PROGRESS.setPrefix("     ");
-    for(int f=0; f<n_count ;f++)
+    
+    // Compute the starting position
+    // -----------------------------------------
+    unsigned int elements = threads_count + 1;
+    unsigned int *Pos = new unsigned int[elements]();
+
+    for (int i = 1; i < threads_count; i++)
     {
-        PROGRESS.inc();
-        if ( isTRK )
-            N = read_fiberTRK( fpTractogram, fiber, n_scalars, n_properties );
+        Pos[i] = Pos[i-1] + batch_size[i];
+    }
+
+    Pos[threads_count] = n_count;
+
+
+    // Check the file extension
+    // -------------------------------------
+    int isTRK;
+
+    char *ext = strrchr(str_filename, '.');
+    if (strcmp(ext,".trk")==0) // for .trk file
+        isTRK = 1;
+    else if (strcmp(ext,".tck")==0) // for .tck file
+        isTRK = 0;
         else
-            N = read_fiberTCK( fpTractogram, fiber , ptrToVOXMM );
-        fiberForwardModel( fiber, N, nReplicas, ptrBlurRho, ptrBlurAngle, ptrBlurWeights, ptrBlurApplyTo[f], ptrHashTable );
+        return 0;    
 
-        kept = 0;
-        if ( FiberSegments.size() > 0 )
-        {
-            if ( FiberLen > minFiberLen && FiberLen < maxFiberLen )
-            {
-                // add segments to files
-                for (it=FiberSegments.begin(); it!=FiberSegments.end(); it++)
-                {
-                    // NB: plese note inverted ordering for 'v'
-                    v = it->first.x + dim.x * ( it->first.y + dim.y * it->first.z );
-                    o = it->first.o;
-                    fwrite( &totFibers,      4, 1, pDict_IC_f );
-                    fwrite( &v,              4, 1, pDict_IC_v );
-                    fwrite( &o,              2, 1, pDict_IC_o );
-                    fwrite( &(it->second),   4, 1, pDict_IC_len );
-                    ptrTDI[ it->first.z + dim.z * ( it->first.y + dim.y * it->first.x ) ] += it->second;
-                    inVoxKey.set( it->first.x, it->first.y, it->first.z );
-                    FiberNorm[inVoxKey] += it->second;
-                }
-                for (fiberNorm=0, itNorm=FiberNorm.begin(); itNorm!=FiberNorm.end(); itNorm++)
-                    fiberNorm += pow(itNorm->second,2);
-                fiberNorm = sqrt(fiberNorm);
-                FiberNorm.clear();
-                fwrite( &fiberNorm,   1, 4, pDict_TRK_norm );   // actual length considered in optimization
-                fwrite( &FiberLen,    1, 4, pDict_TRK_len );    // length of the streamline
-                fwrite( &FiberLenTot, 1, 4, pDict_TRK_lenTot ); // length of the streamline (considering the blur)
-                totICSegments += FiberSegments.size();
-                totFibers++;
-                kept = 1;
+
+    // Open tractogram file and compute the offset for each thread
+    // -----------------------------------------------------------------
+    unsigned long long int current;
+    unsigned long long int OffsetArr[threads_count];
+    int f = 0;
+    float Buff[3];
+    int N;
+
+    FILE* fpTractogram = fopen(str_filename,"rb");
+    if (fpTractogram == NULL) return 0;
+    fseek( fpTractogram, data_offset, SEEK_SET ); // skip the header   
+
+    OffsetArr[0] = ftell( fpTractogram );
+
+    if(isTRK) {
+        while( f != n_count) {
+            fread( (char*)&N, 1, 4, fpTractogram ); // read the number of points in each streamline
+            if( N >= MAX_FIB_LEN || N <= 0 ) return 0;   // check
+            for( int k=0; k<N; k++){
+                fread((char*)Buff, 1, 12, fpTractogram);
             }
+            fseek(fpTractogram,4*n_properties,SEEK_CUR);
+            f++;
+            current = ftell( fpTractogram );
+            for( int i = 1; i < threads_count; i++ ){
+                    if( f == Pos[i] ) 
+                        OffsetArr[i] = current;
+                }
         }
-        fwrite( &kept, 1, 1, pDict_TRK_kept );
+    } else {
+
+        while( f != n_count ) {
+
+            fread((char*)Buff, 1, 12, fpTractogram );
+
+            if( isnan(Buff[0]) ){
+                f++;
+                current = ftell( fpTractogram );
+
+                for( int i = 1; i < threads_count; i++ ){
+                    if( f == Pos[i] ) 
+                        OffsetArr[i] = current;
+                }
+
+            }
+
+        }
+
     }
-    PROGRESS.close();
 
-    fclose( fpTractogram );
-    fclose( pDict_TRK_norm );
-    fclose( pDict_IC_f );
-    fclose( pDict_IC_v );
-    fclose( pDict_IC_o );
-    fclose( pDict_IC_len );
-    fclose( pDict_TRK_len );
-    fclose( pDict_TRK_lenTot );
-    fclose( pDict_TRK_kept );
-
-    printf("     [ %d streamlines kept, %d segments in total ]\n", totFibers, totICSegments );
+    fclose(fpTractogram);
 
 
-    /*=========================*/
-    /*     EC compartments     */
-    /*=========================*/
-    unsigned int totECSegments = 0, totECVoxels = 0;
 
+    // ==========================================
+    //          Parallel IC compartments
+    // ==========================================
+
+    printf( "\n   \033[0;32m* Exporting IC compartments:\033[0m\n" );
+    // unsigned int width = 25;
+    // PROGRESS = new ProgressBar( (unsigned int) n_count, (unsigned int) width);
+    PROGRESS->reset((unsigned int) n_count);
+    PROGRESS->setPrefix("     ");
+    // ---- Original ------
+    for( int i = 0; i<threads_count; i++ ){
+        threads.push_back( thread( ICSegments, str_filename, isTRK, n_count, nReplicas, n_scalars, n_properties, ptrToVOXMM,
+        ptrTDI[i] , ptrBlurRho, ptrBlurAngle, ptrBlurWeights, ptrBlurApplyTo, ptrHashTable, path_out, OffsetArr[i], 
+        i, Pos[i], Pos[i+1]  ) );
+    }
+
+
+    for( int i = 0; i<threads_count; i++ ) {
+        threads[i].join();
+    }
+
+    PROGRESS->close();
+
+    printf( "     [ %d streamlines kept, %d segments in total ]\n", std::accumulate(totFibers.begin(), totFibers.end(), 0), std::accumulate( totICSegments.begin(), totICSegments.end(), 0) );
+    totFibers.clear();
+    threads.clear();
     printf( "\n   \033[0;32m* Exporting EC compartments:\033[0m\n" );
 
-    filename = OUTPUT_path+"/dictionary_EC_v.dict";        FILE* pDict_EC_v   = fopen( filename.c_str(),   "wb" );
-    filename = OUTPUT_path+"/dictionary_EC_o.dict";        FILE* pDict_EC_o   = fopen( filename.c_str(),   "wb" );
+    int EC = ECSegments( ptrPEAKS, Np, vf_THR, ECix, ECiy, ECiz, ptrTDI, ptrHashTable, path_out, ptrPeaksAffine, threads_count );
+
+    printf("     [ %d voxels, %d segments ]\n", totECVoxels, totECSegments );
+
+    return 1;
+
+}
+
+
+int ECSegments(float* ptrPEAKS, int Np, float vf_THR, int ECix, int ECiy, int ECiz,
+    double** ptrTDI, short* ptrHashTable, char* path_out, double* ptrPeaksAffine, int threads){
+
+    // Variables definition
+    string    filename;
+    string    OUTPUT_path(path_out);
+    std::size_t pos = OUTPUT_path.find("/temp");
+    OUTPUT_path = OUTPUT_path.substr (0,pos);
+
+    unsigned short o;
+    unsigned int v;
+    unsigned int temp_totECSegments = 0, temp_totECVoxels = 0;
+
+
+    filename = OUTPUT_path+"/dictionary_EC_v.dict";        FILE* pDict_EC_v       = fopen(filename.c_str(),"wb");
+    filename = OUTPUT_path+"/dictionary_EC_o.dict";        FILE* pDict_EC_o       = fopen(filename.c_str(),"wb");
 
     if ( ptrPEAKS != NULL )
     {
@@ -223,37 +286,44 @@ int trk2dictionary(
         float          norms[ Np ];
         float          *ptr;
         int            ox, oy;
+        int            skip = 0;
 
-        PROGRESS.reset( dim.z );
-        for(iz=0; iz<dim.z ;iz++)
+        for(iz=0; iz<dim.z; iz++)
         {
-            PROGRESS.inc();
-            for(iy=0; iy<dim.y ;iy++)
-            for(ix=0; ix<dim.x ;ix++)
-            {
-                // check if in mask previously computed from IC segments
-                if ( ptrTDI[ iz + dim.z * ( iy + dim.y * ix ) ] == 0 ) continue;
-
-                peakMax = -1;
-                for(id=0; id<Np ;id++)
+            for(iy=0; iy<dim.y; iy++)
+                for(ix=0; ix<dim.x; ix++)
                 {
-                    ptr = ptrPEAKS + 3*(id + Np * ( iz + dim.z * ( iy + dim.y * ix ) ));
-                    dir.x = ptr[0];
-                    dir.y = ptr[1];
-                    dir.z = ptr[2];
-                    norms[id] = dir.norm();
-                    if ( norms[id] > peakMax )
-                        peakMax = norms[id];
-                }
-
-                if ( peakMax > 0 )
-                {
-                    ec_seg.x  = ix;
-                    ec_seg.y  = iy;
-                    ec_seg.z  = iz;
-                    atLeastOne = 0;
+                    // check if in mask previously computed from IC segments
+                    for(int i =0; i<threads; i++){
+                        if ( ptrTDI[i][ iz + dim.z * ( iy + dim.y * ix ) ] == 0 ){
+                            skip += 1;
+                        }
+                    }
+                    if(skip==threads){
+                        skip = 0;
+                        continue;
+                    }
+                    skip = 0;
+                    peakMax = -1;
                     for(id=0; id<Np ;id++)
                     {
+                        ptr = ptrPEAKS + 3*(id + Np * ( iz + dim.z * ( iy + dim.y * ix ) ));
+                        dir.x = ptr[0];
+                        dir.y = ptr[1];
+                        dir.z = ptr[2];
+                        norms[id] = dir.norm();
+                        if ( norms[id] > peakMax )
+                            peakMax = norms[id];
+                    }
+
+                    if ( peakMax > 0 )
+                    {
+                        ec_seg.x  = ix;
+                        ec_seg.y  = iy;
+                        ec_seg.z  = iz;
+                        atLeastOne = 0;
+                        for(id=0; id<Np ;id++)
+                        {
                         if ( norms[id]==0 || norms[id] < vf_THR*peakMax ) continue; // peak too small, don't consider it
 
                         // get the orientation of the current peak
@@ -284,21 +354,155 @@ int trk2dictionary(
                         o = ptrHashTable[ox*181 + oy];
                         fwrite( &v, 4, 1, pDict_EC_v );
                         fwrite( &o, 2, 1, pDict_EC_o );
-                        totECSegments++;
+                        temp_totECSegments++;
                         atLeastOne = 1;
-                    }
+                        }
                     if ( atLeastOne>0 )
-                        totECVoxels++;
+                        temp_totECVoxels++;
+                    }
                 }
-            }
         }
-        PROGRESS.close();
     }
+    totECSegments = temp_totECSegments;
+    totECVoxels = temp_totECVoxels;
 
     fclose( pDict_EC_v );
     fclose( pDict_EC_o );
 
-    printf("     [ %d voxels, %d segments ]\n", totECVoxels, totECSegments );
+    return 1;
+
+}
+
+
+
+
+/********************************************************************************************************************/
+/*                                                Parallel Function                                                 */
+/********************************************************************************************************************/
+
+int ICSegments( char* str_filename, int isTRK, int n_count, int nReplicas, int n_scalars, int n_properties, float* ptrToVOXMM, double* ptrTDI, double* ptrBlurRho, 
+double* ptrBlurAngle, double* ptrBlurWeights, bool* ptrBlurApplyTo, short* ptrHashTable, char* path_out, 
+unsigned long long int offset, int idx, unsigned int startpos, unsigned int endpos ) 
+{
+
+    // Variables definition
+    float           fiber[3][MAX_FIB_LEN] = {0} ;
+    float           fiberNorm;   // normalization
+    unsigned int    N, v, tempTotFibers, temp_totICSegments;
+    unsigned short  o;
+    unsigned char   kept;
+    string          filename;
+    string          OUTPUT_path(path_out);
+
+    unsigned int sumFibers = startpos;
+
+    map<segKey,float>::iterator it;
+    map<segInVoxKey,float> FiberNorm;
+    map<segInVoxKey,float>::iterator itNorm;
+
+    segInVoxKey inVoxKey; 
+
+    P.resize(nReplicas);
+
+
+    // Creo e apro i files per i risultati
+    filename = OUTPUT_path+"/dictionary_TRK_norm_" + std::to_string(idx) + ".dict";   FILE* pDict_TRK_norm = fopen(filename.c_str(),"wb");
+    if ( !pDict_TRK_norm )
+    {
+        printf( "\n[trk2dictionary] Unable to create output files" );
+    }
+    filename = OUTPUT_path+"/dictionary_IC_f_" + std::to_string(idx) + ".dict";        FILE* pDict_IC_f       = fopen(filename.c_str(),"wb");
+    filename = OUTPUT_path+"/dictionary_IC_v_" + std::to_string(idx) + ".dict";        FILE* pDict_IC_v       = fopen(filename.c_str(),"wb");
+    filename = OUTPUT_path+"/dictionary_IC_o_" + std::to_string(idx) + ".dict";        FILE* pDict_IC_o       = fopen(filename.c_str(),"wb");
+    filename = OUTPUT_path+"/dictionary_IC_len_" + std::to_string(idx) + ".dict";      FILE* pDict_IC_len     = fopen(filename.c_str(),"wb");
+    filename = OUTPUT_path+"/dictionary_TRK_len_" + std::to_string(idx) + ".dict";     FILE* pDict_TRK_len    = fopen(filename.c_str(),"wb");
+    filename = OUTPUT_path+"/dictionary_TRK_lenTot_" + std::to_string(idx) + ".dict";  FILE* pDict_TRK_lenTot = fopen(filename.c_str(),"wb");
+    filename = OUTPUT_path+"/dictionary_TRK_kept_" + std::to_string(idx) + ".dict";    FILE* pDict_TRK_kept   = fopen(filename.c_str(),"wb");
+
+
+    // ---- Original -----
+    // Open tractogram file
+    FILE* fpTractogram1 = fopen( str_filename,"rb" );
+    if ( fpTractogram1 == NULL ) return 0; // if there's no tractogram file, then return 0
+    fseek(fpTractogram1, offset, SEEK_SET);
+
+    tempTotFibers = 0;
+    temp_totICSegments = 0;
+    int incr_new = 0;
+    int incr_old = 0;
+    // Iterate over streamlines
+
+    for(int f=startpos; f<endpos; f++) 
+    {        
+
+        if ( isTRK )
+            N = read_fiberTRK( fpTractogram1, fiber, n_scalars, n_properties );
+        else
+            N = read_fiberTCK( fpTractogram1, fiber , ptrToVOXMM );
+
+
+        fiberForwardModel( fiber, N, nReplicas, ptrBlurRho, ptrBlurAngle, ptrBlurWeights, ptrBlurApplyTo[f], ptrHashTable, P );
+
+        kept = 0;
+
+        if ( FiberSegments.size() > 0 )
+        {
+            if ( FiberLen > minFiberLen && FiberLen < maxFiberLen )
+            {
+                // add segments to files
+                for (it=FiberSegments.begin(); it!=FiberSegments.end(); it++)
+                {
+                    // NB: plese note inverted ordering for 'v'
+                    v = it->first.x + dim.x * ( it->first.y + dim.y * it->first.z );
+                    o = it->first.o;       
+
+                    fwrite( &sumFibers,      4, 1, pDict_IC_f );
+                    fwrite( &v,              4, 1, pDict_IC_v );
+                    fwrite( &o,              2, 1, pDict_IC_o );
+                    fwrite( &(it->second),   4, 1, pDict_IC_len );       
+                    
+                    ptrTDI[ it->first.z + dim.z * ( it->first.y + dim.y * it->first.x ) ] += it->second;
+
+                    inVoxKey.set( it->first.x, it->first.y, it->first.z );
+                    FiberNorm[inVoxKey] += it->second;
+                }
+
+                for (fiberNorm=0, itNorm=FiberNorm.begin(); itNorm!=FiberNorm.end(); itNorm++)
+                    fiberNorm += pow(itNorm->second,2);
+                fiberNorm = sqrt(fiberNorm);
+                FiberNorm.clear();
+
+                fwrite( &fiberNorm,   1, 4, pDict_TRK_norm );   // actual length considered in optimization
+                fwrite( &FiberLen,    1, 4, pDict_TRK_len );    // length of the streamline
+                fwrite( &FiberLenTot, 1, 4, pDict_TRK_lenTot ); // length of the streamline (considering the blur)
+
+                temp_totICSegments += FiberSegments.size();
+                sumFibers ++;
+                tempTotFibers ++;
+                kept = 1;
+            }
+
+        }
+
+        fwrite( &kept, 1, 1, pDict_TRK_kept );
+        totFibers[idx] = tempTotFibers;
+        totICSegments[idx] = temp_totICSegments;
+        if (idx == 0){
+            incr_new = std::accumulate(totFibers.begin(), totFibers.end(), 0);
+            for(int i=incr_old; i<incr_new; i++)
+                PROGRESS->inc();
+            incr_old = incr_new;
+        }
+    }
+    fclose( fpTractogram1 );
+    fclose( pDict_TRK_norm );
+    fclose( pDict_IC_f );
+    fclose( pDict_IC_v );
+    fclose( pDict_IC_o );
+    fclose( pDict_IC_len );
+    fclose( pDict_TRK_len );
+    fclose( pDict_TRK_lenTot );
+    fclose( pDict_TRK_kept );
 
     return 1;
 }
@@ -307,12 +511,12 @@ int trk2dictionary(
 /********************************************************************************************************************/
 /*                                                 fiberForwardModel                                                */
 /********************************************************************************************************************/
-void fiberForwardModel( float fiber[3][MAX_FIB_LEN], unsigned int pts, int nReplicas, double* ptrBlurRho, double* ptrBlurAngle, double* ptrBlurWeights, bool doApplyBlur, short* ptrHashTable )
+void fiberForwardModel( float fiber[3][MAX_FIB_LEN], unsigned int pts, int nReplicas, double* ptrBlurRho, double* ptrBlurAngle, double* ptrBlurWeights, bool doApplyBlur, short* ptrHashTable, vector<Vector<double>>& P )
 {
-    static Vector<double> S1, S2, S1m, S2m, P_old, P_int, q, n, nr, qxn, qxqxn;
-    static Vector<double> vox, vmin, vmax, dir1, dir2;
-    static double         len, t, alpha, w, R, dot;
-    static int            i, j, k;
+    thread_local static Vector<double> S1, S2, S1m, S2m, P_old, P_int, q, n, nr, qxn, qxqxn;
+    thread_local static Vector<double> vox, vmin, vmax, dir1, dir2;
+    thread_local static double         len, t, alpha, w, R, dot;
+    thread_local static int            i, j, k;
 
     FiberLen = 0.0;
     FiberLenTot = 0.0;
@@ -486,11 +690,11 @@ void fiberForwardModel( float fiber[3][MAX_FIB_LEN], unsigned int pts, int nRepl
 /********************************************************************************************************************/
 void segmentForwardModel( const Vector<double>& P1, const Vector<double>& P2, int k, double w, short* ptrHashTable )
 {
-    static Vector<int>    vox;
-    static Vector<double> dir, dirTrue;
-    static double         longitude, colatitude, len;
-    static segKey         key;
-    static int            ox, oy;
+    thread_local static Vector<int>    vox;
+    thread_local static Vector<double> dir, dirTrue;
+    thread_local static double         longitude, colatitude, len;
+    thread_local static segKey         key;
+    thread_local static int            ox, oy;
 
     // direction of the segment
     dir.y = P2.y-P1.y;
@@ -539,8 +743,8 @@ void segmentForwardModel( const Vector<double>& P1, const Vector<double>& P2, in
 /********************************************************************************************************************/
 bool rayBoxIntersection( Vector<double>& origin, Vector<double>& direction, Vector<double>& vmin, Vector<double>& vmax, double & t)
 {
-    static double tmin, tmax, tymin, tymax, tzmin, tzmax;
-    static Vector<double> invrd;
+    thread_local static double tmin, tmax, tymin, tymax, tzmin, tzmax;
+    thread_local static Vector<double> invrd;
 
     // inverse direction to catch float problems
     invrd.x = 1.0 / direction.x;
@@ -622,7 +826,7 @@ unsigned int read_fiberTRK( FILE* fp, float fiber[3][MAX_FIB_LEN], int ns, int n
 unsigned int read_fiberTCK( FILE* fp, float fiber[3][MAX_FIB_LEN], float* ptrToVOXMM )
 {
     int i = 0;
-    float P[3];
+    thread_local float P[3];
     fread((char*)P, 1, 12, fp);
     while( !(isnan(P[0])) && !(isnan(P[1])) &&  !(isnan(P[2])) )
     {